avances y maleficios en la practica 3

+++ Practicas/Practica3/unsharp_masking.m
@@ -1,10 +1,10 @@
 function sharp = unsharp_masking(img, mascara, const)
 
-pasa_bajo = conv2(img, mascara,"same");
+pasa_bajo = uint8(conv2(img, mascara,"same"));
 
 pasa_alto = img - pasa_bajo;
 
-sharp = img + pasa_alto * const;
+sharp = uint8(img + pasa_alto * const);
 
 
 endfunction
